+++ src/overloaded/Overloaded.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+// helper type for std::visit
+template<typename... T>
+struct Overloaded : T...
+{
+    using T::operator()...;
+};
+
+template<class... T>
+Overloaded(T...) -> Overloaded<T...>;
